+++ frontend/lib/utils/search-data.ts
@@ -131,20 +131,22 @@ export const searchData: SearchItem[] = [
     keywords: ['origin', '源站', '后端', 'backend', '服务器', '负载均衡'],
   },
   {
-    id: 'console-error-pages',
-    title: '错误页',
-    description: '配置源站错误状态码触发的统一 HTML 错误页',
-    url: '/error-pages',
+    id: 'console-responses',
+    title: '响应页面',
+    description: '配置源站错误页与离线兜底联系页',
+    url: '/responses',
     category: 'page',
     keywords: [
+      '响应页面',
       '错误页',
+      '联系页',
       '源站',
       '502',
       '503',
       '5xx',
+      'Service Worker',
+      '离线兜底',
       'error page',
-      'origin error',
-      'error_page',
     ],
   },
   {
